Split ReportList fragments to separate classes
