+++ app/src/main/java/lt/vilnius/tvarkau/dagger/component/ApplicationComponent.java
@@ -10,12 +10,12 @@ import lt.vilnius.tvarkau.dagger.module.AnalyticsModule;
 import lt.vilnius.tvarkau.dagger.module.DataModule;
 import lt.vilnius.tvarkau.dagger.module.LegacyApiModule;
 import lt.vilnius.tvarkau.dagger.module.SharedPreferencesModule;
+import lt.vilnius.tvarkau.fragments.AllReportsListFragment;
 import lt.vilnius.tvarkau.fragments.BaseFragment;
 import lt.vilnius.tvarkau.fragments.BaseMapFragment;
 import lt.vilnius.tvarkau.fragments.MultipleProblemsMapFragment;
 import lt.vilnius.tvarkau.fragments.NewReportFragment;
 import lt.vilnius.tvarkau.fragments.PhotoInstructionsFragment;
-import lt.vilnius.tvarkau.fragments.ProblemsListFragment;
 import lt.vilnius.tvarkau.fragments.ReportFilterFragment;
 import lt.vilnius.tvarkau.fragments.ReportImportDialogFragment;
 import lt.vilnius.tvarkau.fragments.ReportTypeListFragment;
@@ -52,5 +52,5 @@ public interface ApplicationComponent {
 
     void inject(ReportFilterFragment fragment);
 
-    void inject(ProblemsListFragment fragment);
+    void inject(AllReportsListFragment fragment);
 }
