+++ app/src/main/java/lt/vilnius/tvarkau/fragments/interactors/AllReportListInteractor.kt
@@ -38,7 +38,7 @@ class AllReportListInteractor(
 
         return legacyApiService.getProblems(GetReportListRequest(params))
                 .subscribeOn(ioScheduler)
-                .map { it.result }
+                .map { it.result ?: emptyList() }
                 .toSingle()
     }
