+++ app/src/main/res/layout/fragment_my_reports_list.xml
@@ -1,26 +1,9 @@
 <android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <android.support.v4.widget.SwipeRefreshLayout
-        android:id="@+id/swipe_container"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <android.support.v7.widget.RecyclerView
-            android:id="@+id/problem_list"
-            android:name="lt.vilnius.tvarkau.ProblemListFragment"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="#fff"
-            android:scrollbars="vertical"
-            app:layoutManager="LinearLayoutManager"
-            tools:context=".ProblemsListActivity"
-            tools:listitem="@layout/problem_list_content" />
-
-    </android.support.v4.widget.SwipeRefreshLayout>
+    <include layout="@layout/include_report_list_recycler_view" />
 
     <LinearLayout
         android:id="@+id/my_problems_empty_view"
@@ -78,11 +61,6 @@
                 android:text="@string/my_problems_import_from_vilnius_account"
                 android:textColor="@color/colorPrimary" />
         </LinearLayout>
-
     </LinearLayout>
-
-    <include layout="@layout/no_internet" />
-
-    <include layout="@layout/server_not_responding" />
 
 </android.support.design.widget.CoordinatorLayout>
