+++ app/src/main/res/layout/include_report_list_recycler_view.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.v4.widget.SwipeRefreshLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/swipe_container"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:showIn="@layout/fragment_all_reports_list">
+
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/report_list"
+        android:name="lt.vilnius.tvarkau.ProblemListFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="#fff"
+        android:scrollbars="vertical"
+        app:layoutManager="LinearLayoutManager"
+        tools:context=".ProblemsListActivity"
+        tools:listitem="@layout/problem_list_content" />
+</android.support.v4.widget.SwipeRefreshLayout>
