+++ app/src/test/java/lt/vilnius/tvarkau/fragments/presenters/AllReportsListPresenterImplTest.kt
@@ -6,6 +6,7 @@ import lt.vilnius.tvarkau.fragments.interactors.ReportListInteractor
 import lt.vilnius.tvarkau.fragments.presenters.ConnectivityProviderImpl.NetworkConnectivityError
 import lt.vilnius.tvarkau.fragments.views.ReportListView
 import org.junit.Test
+import rx.Single
 import rx.Single.error
 import rx.Single.just
 import rx.schedulers.Schedulers
@@ -34,32 +35,42 @@ class AllReportsListPresenterImplTest {
     fun loadReports_success() {
         whenever(interactor.getProblems(any())).thenReturn(just(reports))
 
-        fixture.getReportsForPage(1)
+        fixture.getReportsForPage(0)
 
-        verify(view).markLoading(true)
-        verify(view).markLoading(false)
+        verify(view).showProgress()
+        verify(view).hideProgress()
         verify(view).onReportsLoaded(reports)
+    }
+
+    @Test
+    fun loadReports_noResults_showEmptyState() {
+        whenever(interactor.getProblems(any())).thenReturn(just(emptyList()))
+
+        fixture.getReportsForPage(0)
+
+        verify(view).showEmptyState()
+        verify(view, never()).onReportsLoaded(any())
     }
 
     @Test
     fun loadReports_noNetwork_showNetworkError() {
         whenever(connectivityProvider.ensureConnected()).thenReturn(error(NetworkConnectivityError()))
 
-        fixture.getReportsForPage(1)
+        fixture.getReportsForPage(0)
 
-        verify(view).showNetworkError(1)
+        verify(view).showNetworkError(0)
         verify(interactor, never()).getProblems(any())
     }
 
     @Test
     fun loadReports_otherError_showError() {
         val exception = RuntimeException("Some error")
-        whenever(interactor.getProblems(1)).thenReturn(error(exception))
+        whenever(interactor.getProblems(0)).thenReturn(Single.error(exception))
 
-        fixture.getReportsForPage(1)
+        fixture.getReportsForPage(0)
 
-        verify(view).markLoading(true)
-        verify(view).markLoading(false)
+        verify(view).showProgress()
+        verify(view).hideProgress()
         verify(view).showError(exception)
     }
 }
